requirements.txt
streamlit
python-dotenv
together